requirements.txt
progressbar2
PyPDF4